relocating the pie logic
