+++ frontend/src/pocketfy/pages/PocketfyPage/components/PieChart.jsx
@@ -1,9 +1,40 @@
+import { useMemo } from 'react';
 import { Chart as ChartJS, ArcElement, Tooltip, Legend } from 'chart.js';
 import { Pie } from 'react-chartjs-2';
 
 ChartJS.register(ArcElement, Tooltip, Legend);
 
-export const PieChart = ({data}) => {
+export const PieChart = ({ categoriesAmounts = {}, type }) => {
+
+  const data = useMemo(() => {
+    return {
+      labels: Object.keys(categoriesAmounts[type]),
+      datasets: [
+        {
+          label: 'Amount',
+          data: Object.values(categoriesAmounts[type]),
+          backgroundColor: [
+            'rgba(255, 99, 132, 0.2)',
+            'rgba(54, 162, 235, 0.2)',
+            'rgba(255, 206, 86, 0.2)',
+            'rgba(75, 192, 192, 0.2)',
+            'rgba(153, 102, 255, 0.2)',
+            'rgba(255, 159, 64, 0.2)',
+          ],
+          borderColor: [
+            'rgba(255, 99, 132, 1)',
+            'rgba(54, 162, 235, 1)',
+            'rgba(255, 206, 86, 1)',
+            'rgba(75, 192, 192, 1)',
+            'rgba(153, 102, 255, 1)',
+            'rgba(255, 159, 64, 1)',
+          ],
+          borderWidth: 1,
+        }
+      ]
+    }
+  }, [categoriesAmounts]);
+
   return (
     <Pie data={data} />
   )
